+++ test/Models/Catalog/Ows/WebMapServiceCatalogItemSpec.ts
@@ -43,11 +43,7 @@ describe("WebMapServiceCatalogItem", function() {
     const wms = new WebMapServiceCatalogItem("test", terria);
     runInAction(() => {
       wms.setTrait("definition", "url", "test/WMS/single_metadata_url.xml");
-      wms.setTrait(
-        "definition",
-        "layers",
-        "mobile-black-spot-programme:funded-base-stations-group"
-      );
+      wms.setTrait("definition", "layers", "single_period");
     });
     return wms.loadMapItems();
   });
@@ -128,11 +124,7 @@ describe("WebMapServiceCatalogItem", function() {
     wms = new WebMapServiceCatalogItem("test", terria);
     runInAction(() => {
       wms.setTrait("definition", "url", "test/WMS/single_metadata_url.xml");
-      wms.setTrait(
-        "definition",
-        "layers",
-        "mobile-black-spot-programme:funded-base-stations-group"
-      );
+      wms.setTrait("definition", "layers", "single_period");
     });
     let mapItems: ImageryParts[] = [];
     const cleanup = autorun(() => {
@@ -190,7 +182,7 @@ describe("WebMapServiceCatalogItem", function() {
     runInAction(() => {
       wms.setTrait("definition", "url", "test/WMS/wms_1_1_1.xml");
       wms.setTrait("definition", "useWmsVersion130", false);
-      wms.setTrait("definition", "layers", "GA_Topo_10M");
+      wms.setTrait("definition", "layers", "IDZ10004");
     });
     let mapItems: ImageryParts[] = [];
     const cleanup = autorun(() => {
@@ -314,11 +306,7 @@ describe("WebMapServiceCatalogItem", function() {
     wms = new WebMapServiceCatalogItem("test", terria);
     runInAction(() => {
       wms.setTrait("definition", "url", "test/WMS/single_metadata_url.xml");
-      wms.setTrait(
-        "definition",
-        "layers",
-        "mobile-black-spot-programme:funded-base-stations-group"
-      );
+      wms.setTrait("definition", "layers", "single_period");
       wms.setTrait("definition", "tileWidth", 512);
       wms.setTrait("definition", "tileHeight", 512);
     });
@@ -338,6 +326,61 @@ describe("WebMapServiceCatalogItem", function() {
     } finally {
       cleanup();
     }
+  });
+
+  it("uses query parameters from URL - no parameters", async function() {
+    let wms: WebMapServiceCatalogItem;
+    const terria = new Terria();
+    wms = new WebMapServiceCatalogItem("test", terria);
+    runInAction(() => {
+      wms.setTrait("definition", "url", "test/WMS/single_metadata_url.xml");
+    });
+
+    await wms.loadMetadata();
+
+    expect(wms.tileHeight).toBe(256);
+    expect(wms.tileWidth).toBe(256);
+    expect(wms.layers).toBe("single_period");
+    expect(wms.styles).toBeUndefined();
+    expect(wms.useWmsVersion130).toBeTruthy();
+    expect(wms.crs).toBe("EPSG:3857");
+  });
+
+  it("uses query parameters from URL - with parameters", async function() {
+    let wms: WebMapServiceCatalogItem;
+    const terria = new Terria();
+    wms = new WebMapServiceCatalogItem("test", terria);
+    runInAction(() => {
+      wms.setTrait(
+        "definition",
+        "url",
+        "test/WMS/single_metadata_url.xml?&styles=jet&version=1.1.1&crs=EPSG%3A4326&service=WMS&request=GetCapabilities&layers=single_period&width=512&height=512"
+      );
+    });
+
+    await wms.loadMetadata();
+
+    expect(wms.tileHeight).toBe(512);
+    expect(wms.tileWidth).toBe(512);
+    expect(wms.layers).toBe("single_period");
+    expect(wms.styles).toBe("jet");
+    expect(wms.useWmsVersion130).toBeFalsy();
+    expect(wms.crs).toBe("EPSG:4326");
+  });
+
+  it("invalid/valid layers", async function() {
+    let wms: WebMapServiceCatalogItem;
+    const terria = new Terria();
+    wms = new WebMapServiceCatalogItem("test", terria);
+    runInAction(() => {
+      wms.setTrait("definition", "url", "test/WMS/single_metadata_url.xml");
+      wms.setTrait("definition", "layers", "invalidLayer,single_period");
+    });
+
+    await wms.loadMetadata();
+
+    expect(wms.invalidLayers).toEqual(["invalidLayer"]);
+    expect(wms.validLayers).toEqual(["single_period"]);
   });
 
   it("uses GetFeatureInfo from GetCapabilities", async function() {