+++ CHANGES.md
@@ -11,6 +11,7 @@ Change Log
 * Add `legendBackgroundColor` to `LegendOwnerTraits` and `backgroundColor` to `LegendTraits`
 * Filter `"styles","version","format","srs","crs"` conflicting query parameters from WMS `url`
 * WMS `styles`, `tileWidth`, `tileHeight` and `crs`/`srs` will use value in `url` if it is defined (similar to existing behavior with `layers`)
+* WMS will now show warning if invalid `layers` (eg if the specified `layers` don't exist in `GetCapabilities`)
 * [The next improvement]
 
 #### 8.1.24 - 2022-03-08